database-controller/requirements.txt
Flask[async]==2.2.5 # Updating to 3.0+ currently gives errors so temporarily setting to older version
gunicorn==23.0.0
Werkzeug==3.0.6
flask_cors==3.0.10
requests==2.31.0
prisma==0.15.0
python-dotenv==1.0.1
kubernetes==29.0.0
asyncio
passlib
bcrypt
PyJWT
jsonschema
